% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_db_path.R
\name{get_default_db_path}
\alias{get_default_db_path}
\title{Get Default Database Path}
\usage{
get_default_db_path()
}
\value{
Path to the default SQLite database
}
\description{
Get Default Database Path
}
