% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch Receipt Tracker App}
\usage{
launch_app()
}
\arguments{
\item{...}{Additional parameters passed to shinyApp}
}
\value{
A Shiny app object
}
\description{
Launch Receipt Tracker App
}
