% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_receipt_data.R
\name{load_receipt_data}
\alias{load_receipt_data}
\title{Load All Receipt Data}
\usage{
load_receipt_data(db_path = get_default_db_path())
}
\arguments{
\item{db_path}{Path to SQLite database file}
}
\value{
Data frame with all stored receipt data
}
\description{
Load All Receipt Data
}
