% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_table.R
\name{mod_data_table_server}
\alias{mod_data_table_server}
\title{Data Table Module Server}
\usage{
mod_data_table_server(id, receipt_data)
}
\arguments{
\item{id}{The module identifier}

\item{receipt_data}{Reactive containing receipt data}
}
\value{
A module server function
}
\description{
Data Table Module Server
}
