% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_table.R
\name{mod_data_table_ui}
\alias{mod_data_table_ui}
\title{Data Table Module UI}
\usage{
mod_data_table_ui(id)
}
\arguments{
\item{id}{The module identifier}
}
\value{
A shiny UI component
}
\description{
Data Table Module UI
}
