% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_receipt_history.R
\name{mod_receipt_history_server}
\alias{mod_receipt_history_server}
\title{Receipt History Module Server}
\usage{
mod_receipt_history_server(id, receipt_data, historical_data)
}
\arguments{
\item{id}{The module identifier}

\item{receipt_data}{Reactive containing receipt data}

\item{historical_data}{Reactive containing all historical receipt data}
}
\value{
A module server function
}
\description{
Receipt History Module Server
}
