% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_receipt_history.R
\name{mod_receipt_history_ui}
\alias{mod_receipt_history_ui}
\title{Receipt History Module UI}
\usage{
mod_receipt_history_ui(id)
}
\arguments{
\item{id}{The module identifier}
}
\value{
A shiny UI component
}
\description{
Receipt History Module UI
}
