% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_receipt_upload.R
\name{mod_receipt_upload_server}
\alias{mod_receipt_upload_server}
\title{Receipt Upload Module Server}
\usage{
mod_receipt_upload_server(id, receipt_data)
}
\arguments{
\item{id}{The module identifier}

\item{receipt_data}{Reactive to store processed receipt data}
}
\value{
A module server function
}
\description{
Receipt Upload Module Server
}
