% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_receipt_upload.R
\name{mod_receipt_upload_ui}
\alias{mod_receipt_upload_ui}
\title{Receipt Upload Module UI}
\usage{
mod_receipt_upload_ui(id)
}
\arguments{
\item{id}{The module identifier}
}
\value{
A shiny UI component
}
\description{
Receipt Upload Module UI
}
