% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_receipt_data.R
\name{save_receipt_data}
\alias{save_receipt_data}
\title{Save Receipt Data to Database}
\usage{
save_receipt_data(data, db_path = get_default_db_path())
}
\arguments{
\item{data}{Data frame with receipt data}

\item{db_path}{Path to SQLite database file}
}
\value{
Logical indicating success
}
\description{
Save Receipt Data to Database
}
